+++ src/types/index.ts
@@ -4,4 +4,5 @@ export * from './Auth.js';
 export * from './Callbacks.js';
 export * from './DomainRecordList.js';
 export * from './Record.js';
+export * from './Options.js';
 export * from './ZoneMap.js';